разделили один app на несколько файлов
