+++ SysCall/config.py
@@ -1,0 +1,6 @@
+
+BEACONS_FILE = "standart.beacons"
+
+MQTT_BROKER = "localhost"
+MQTT_PORT = 1883
+MQTT_TOPIC = "registrar/data"
